% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_asserts.R
\name{assert_df}
\alias{assert_df}
\title{Check data frame and columns}
\usage{
assert_df(df, cols)
}
\arguments{
\item{df}{A data frame to check}

\item{cols}{Character vector of columns to check for}
}
\description{
Check data frame and columns
}
