% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_asserts.R
\name{assert_df_cols}
\alias{assert_df_cols}
\title{Check that columns are in a data frame}
\usage{
assert_df_cols(df, cols)
}
\arguments{
\item{df}{A data frame to check}

\item{cols}{Character vector of columns to check for}
}
\description{
Check that columns are in a data frame
}
