% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_asserts.R
\name{assert_df_type}
\alias{assert_df_type}
\title{Check that a data frame is a data frame}
\usage{
assert_df_type(df)
}
\arguments{
\item{df}{A data frame to check}
}
\description{
Check that a data frame is a data frame
}
