% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_asserts.R
\name{assert_pd}
\alias{assert_pd}
\title{Check that parsed dataframe list is passed correctly}
\usage{
assert_pd(pd)
}
\arguments{
\item{pd}{Parsed dataframe list to check}
}
\description{
Check that parsed dataframe list is passed correctly
}
