% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{database_names}
\alias{database_names}
\title{Names of columns in data frames parsed from the FFP databases.}
\format{
A data frame with 134 rows and 4 variables:
\describe{
\item{table}{Table that the name is found in. Either "session", "attendance", and "attendees".}
\item{col_names}{Original names found in the FFP databases.}
\item{col_codes}{Corresponding column codes for use when parsing FFP database files.}
}
}
\usage{
database_names
}
\description{
Contains all possible names found in FFP parsed databases from the UK
and Jamaica. The file is used to standardize R data frame names so that data
from each database can be pulled into a single data frame.
}
\keyword{datasets}
