% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_attendance.R
\name{get_full_session_df}
\alias{get_full_session_df}
\title{Get data frame with all session and attendance data}
\usage{
get_full_session_df(pd)
}
\arguments{
\item{pd}{Parsed database list coming out of \code{\link[=parse_session_database]{parse_session_database()}} or
\code{\link[=parse_multiple_session_databases]{parse_multiple_session_databases()}} with \code{sessions}, \code{attendance},  and \code{attendees}
data frames.}
}
\value{
A data frame.
}
\description{
\code{get_full_session_df()} pulls together the separate \code{sessions}, \code{attendance}, and
\code{attendees} data frames into a single data frame with 1 row for each attendee
attending a session. Useful to get a single combined database of all attendance
data to later parse for dashboard visualizations or other analysis.
}
