% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{get_pbi_dashboard_data}
\alias{get_pbi_dashboard_data}
\title{Create files necessary for FFP Power BI dashboard}
\usage{
get_pbi_dashboard_data(pd, file = NULL, ...)
}
\arguments{
\item{pd}{Parsed database list coming out of \code{\link[=parse_single_database]{parse_single_database()}} or
\code{\link[=parse_multiple_databases]{parse_multiple_databases()}} with \code{sessions}, \code{attendance},  and \code{attendees}
data frames.}

\item{file}{If not \code{NULL}, file name and path for writing, must end in \code{".xlsx"}.
If \code{NULL}, no Excel file is generated.}

\item{...}{Additional arguments passed to \code{\link[openxlsx:write.xlsx]{openxlsx::write.xlsx()}}.}
}
\value{
A list of data frames, and Excel file printed to \code{file}.
}
\description{
\code{get_pbi_dashboard_data()} takes in the parsed databases coming from
\code{\link[=parse_single_database]{parse_single_database()}} or \code{\link[=parse_multiple_databases]{parse_multiple_databases()}} and produces the
data frames necessary for use in the Power BI dashboards. The function exports
a list of data frames that will go into the dashboards in R, and also exports
them in Excel to a specified file path.
}
