% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_london.R
\name{get_pbi_dashboard_london}
\alias{get_pbi_dashboard_london}
\title{Create files necessary for FFP London's Power BI dashboard}
\usage{
get_pbi_dashboard_london(pd, file = NULL, ...)
}
\arguments{
\item{pd}{Parsed database list coming out of \code{\link[=parse_session_database]{parse_session_database()}} or
\code{\link[=parse_multiple_session_databases]{parse_multiple_session_databases()}} with \code{sessions}, \code{attendance},  and \code{attendees}
data frames.}

\item{file}{If not \code{NULL}, file name and path for writing, must end in \code{".xlsx"}.
If \code{NULL}, no Excel file is generated.}

\item{...}{Additional arguments passed to \code{\link[openxlsx:write.xlsx]{openxlsx::write.xlsx()}}.}
}
\value{
A list of data frames, and Excel file printed to \code{file}.
}
\description{
\code{get_pbi_dashboard_london()} takes in the parsed databases coming from
\code{\link[=parse_session_database]{parse_session_database()}} or \code{\link[=parse_multiple_session_databases]{parse_multiple_session_databases()}} and produces the
data frames necessary for use in FFP London's Power BI dashboards.
The function exports a list of data frames that will go into the dashboards in R,
and also exports them in Excel to a specified file path.
}
\details{
The data frames filter data based on a specific set of attributes. Specifically:
\itemize{
\item Open Access filters data for sessions where \code{project == "Open Access"}
\item Employability filters data for sessions where \code{project == "Employment"}
\item Pathways filters data for sessions where \code{activity_group == "Pathways"}
\item Special Project filters data for sessions where \code{project == "Special Project"}
\item Youth leadership filters data for sessions where \code{project == "Youth Leadership"}
}

Relevant columns are then selected for each of these activity groups, written
to Excel, and exported as a list.
}
