% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_quarter.R
\name{get_quarter}
\alias{get_quarter}
\title{Get quarter from date vector}
\usage{
get_quarter(x)
}
\arguments{
\item{x}{Date vector.}
}
\value{
Character vector.
}
\description{
\code{get_quarter()} takes a date vector and gets quarter from it in the format
\verb{YYYY QD}, such as \verb{2020 Q1}.
}
