% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_jamaica.R
\name{jamaica_pbi_df}
\alias{jamaica_pbi_df}
\title{Get data for the Jamaica PBI dashboard}
\usage{
jamaica_pbi_df(df)
}
\arguments{
\item{df}{Full data frame of attendance and session data produced by \code{\link[=get_full_df]{get_full_df()}}.}
}
\value{
A data frame.
}
\description{
\code{jamaica_pbi_df()} creates a data frame with the columns and values necessary
for providing to the Open Access dashboard in Power BI. This is generated on
all rows within the Jamaica data provided.
}
