% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_headcount.R
\name{parse_headcount_database}
\alias{parse_headcount_database}
\title{Parse Excel export of headcount data from the Fight for Peace database}
\usage{
parse_headcount_database(path)
}
\arguments{
\item{path}{File path to Excel file}
}
\value{
A data frame.
}
\description{
\code{parse_headcount_database()} parses the Fight for Peace headcount database export (from Excel)
format into a single frame.
}
