% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_databases.R
\name{parse_multiple_databases}
\alias{parse_multiple_databases}
\title{Load multiple FFP databases}
\usage{
parse_multiple_databases(paths)
}
\arguments{
\item{paths}{File paths to FFP databases}
}
\description{
Load multiple FFP databases
}
