% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_databases.R
\name{parse_multiple_headcount_databases}
\alias{parse_multiple_headcount_databases}
\title{Load multiple FFP headcount databases}
\usage{
parse_multiple_headcount_databases(paths)
}
\arguments{
\item{paths}{File paths to FFP databases}
}
\value{
A data frame.
}
\description{
\code{parse_multiple_headcount_databases()} loads in multiple Excel files containing data
exported from the FFP headcount database and combines them together. This can be used
to combine databases from different country offices or even from different years.
}
