% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.fight_prediction}
\alias{print.fight_prediction}
\title{Print Method for Fight Predictions}
\usage{
\method{print}{fight_prediction}(x, ...)
}
\arguments{
\item{x}{An object of class \code{fight_prediction}.}

\item{...}{Additional arguments passed to print.}
}
\description{
Custom print method to format the output of model_fight_outcome cleanly to the console using the cli package.
}
