% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{special_project_df}
\alias{special_project_df}
\title{Get data for the Special Project dashboard}
\usage{
special_project_df(df)
}
\arguments{
\item{df}{Full data frame of attendance and session data produced by \code{\link[=get_full_df]{get_full_df()}}.}
}
\value{
A data frame.
}
\description{
\code{special_project_df()} creates a data frame with the columns and values necessary
for providing to the Special Project dashboard in Power BI. This is generated by
filtering for \code{project == "Special Project"} and selecting relevant columns.
}
