% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_london.R
\name{youth_leadership_df}
\alias{youth_leadership_df}
\title{Get data for the Youth Leadership dashboard}
\usage{
youth_leadership_df(df)
}
\arguments{
\item{df}{Full data frame of attendance and session data produced by \code{\link[=get_full_session_df]{get_full_session_df()}}.}
}
\value{
A data frame.
}
\description{
\code{youth_leadership_df()} creates a data frame with the columns and values necessary
for providing to the Youth Leadership dashboard in Power BI. This is generated by
filtering for \code{project == "Youth Leadership"} and selecting relevant columns.
}
